+++ test/test_environment/test_battlefield_wrapper.py
@@ -66,7 +66,7 @@ class TestBattleFieldWrapper(unittest.TestCase):
 
         # Create a mock state with shape (L, L, 36)
         L = 10  # Map size
-        state = np.zeros((L, L, 37), dtype=np.int8)
+        state = np.zeros((L, L, 37), dtype=np.float16)
 
         # Define channel indices
         OBSTACLE_CHANNEL = 0
